Moving components out for reuse
